Rename sampler directory to proposal.
