+++ test/integration/test_HMC.py
@@ -1,4 +1,4 @@
-from flowMC.sampler.HMC import HMC
+from flowMC.proposal.HMC import HMC
 import jax
 import jax.numpy as jnp
 from jax.scipy.special import logsumexp
@@ -59,7 +59,7 @@ state = HMC_sampler.sample(subkey, n_local_steps, initial_position[:, 0], data)
 
 
 from flowMC.nfmodel.rqSpline import MaskedCouplingRQSpline
-from flowMC.sampler.Sampler import Sampler
+from Sampler import Sampler
 
 n_dim = 5
 n_chains = 2
