+++ test/integration/test_MALA.py
@@ -1,4 +1,4 @@
-from flowMC.sampler.MALA import MALA
+from flowMC.proposal.MALA import MALA
 import jax
 import jax.numpy as jnp
 from jax.scipy.special import logsumexp
@@ -62,7 +62,7 @@ state = MALA_Sampler.sample(subkey, n_local_steps, initial_position[:, 0], data)
 
 
 from flowMC.nfmodel.rqSpline import MaskedCouplingRQSpline
-from flowMC.sampler.Sampler import Sampler
+from Sampler import Sampler
 
 n_dim = 5
 n_chains = 2
