+++ test/integration/test_flowHMC.py
@@ -1,11 +1,11 @@
-from flowMC.sampler.MALA import MALA
-from flowMC.sampler.flowHMC import flowHMC
+from flowMC.proposal.MALA import MALA
+from flowMC.proposal.flowHMC import flowHMC
 import jax
 import jax.numpy as jnp
 from jaxtyping import Float, Array
 from jax.scipy.special import logsumexp
 from flowMC.nfmodel.rqSpline import MaskedCouplingRQSpline
-from flowMC.sampler.Sampler import Sampler
+from Sampler import Sampler
 
 
 def log_posterior(x, data):
